ai-services/certificate_detector/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pytesseract==0.3.13
pillow==10.4.0
opencv-python-headless==4.10.0.84
pypdf2==3.0.1
numpy==1.26.4